Reorganised code creating various functions
Code was originally one continuous piece, with no callable functions
which led to the code not only being confusing to read, but also
limiting usage cases and making further expansion of the code
difficult.

It is in a working state for now, and is in a good starting place for
further optimising the code and adding new features and making it more
user friendly.