Allows users to choose prog length and chord freq
Progression length and chord frequency can now be chosen by the user
with values from 1-99 being supported. Whilst this is a very wide range
I'd like to keep it this way for now, until I figure out a more
sensible range.

There is possibly some redundant code which needs to be addressed, but
this is not urgent. Closes #5