Add option to allow generator to use uncommon modes
There is now an option after initializing the code that will ask the
user to input 1 to allow uncommon modes to be used. Whilst this
implementation isn't ideal, optimising this is something I plan to do
see issue #13.

For now I will consider issue #4 closed, but will refrain from merging
onto main branch as the implementation is very messy. The issue will be
manually closed on GitHub.