Complete overhaul of user interaction
No longer is the user prompted to input values for every variable in
the script, they are now optional arguments which can be appended to
the execution of the script. If the user does not wish to do so default
values are used instead.

This closes #13 and #14. I've left the old code in for now just
commented. This is mainly since there are currently no safeguards to
stop the user from inputting unwanted values, so this might be useful
for future reference.